// packages/astro/templates/default/src/data/types.ts
export type AppScores = {
  activity: number;
  maturity: number;
  learning: number;
  contribution: number;
  docs: number;
  overall: number;
};